detach uart2 with optional buffer

+++ uCNC/src/utils.h
@@ -434,7 +434,7 @@ extern "C"
 
 #define __TIMEOUT_US__(timeout) for (int32_t elap_us_##timeout, curr_us_##timeout = mcu_free_micros(); ((int32_t)timeout) >= 0; elap_us_##timeout = mcu_free_micros() - curr_us_##timeout, timeout -= ABS(elap_us_##timeout), curr_us_##timeout = mcu_free_micros())
 #define __TIMEOUT_MS__(timeout) timeout*=1000; __TIMEOUT_US__(timeout)
-#define __TIMEOUT_ASSERT__(timeout) if(timeout < 0)
+#define __TIMEOUT_ASSERT__(timeout) if(((int32_t)timeout) < 0)
 
 #if defined(__GNUC__) && __GNUC__ >= 7
  #define __FALL_THROUGH__ __attribute__ ((fallthrough));
